win32: refactor nix_mx dataarray
